pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TEMdb"
version = "0.1.0"
authors = [
    { name="Derrick Brittain", email="derrickb@alleninstitute.org" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "beanie",
    "pydantic",
    "pydantic-settings",
    "jinja2",
    "python-multipart",
    "uvicorn-worker"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",      
    "httpx",     
    "faker",           
    "mongomock-motor", 
]
docs = [
    "mkdocs", 
    "mkdocs-material",
    "mkdocs-mermaid2-plugin",
    "pymdown-extensions",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]


[tool.setuptools]
py-modules = ["temdb"]